+++ src/com/tomclaw/mandarin/xmpp/GroupChatNickEditFrame.java
@@ -37,7 +37,7 @@ public class GroupChatNickEditFrame extends Window {
 
       public void actionPerformed() {
         try {
-          XmppSender.sendPresence( xmppAccountRoot.xmppSession.xmlWriter, null, xmppItem.jid.concat( "/" ).concat( nickField.getText() ), null, XmppStatusUtil.getStatusDescr( xmppAccountRoot.statusId ), null, 5, false, null, null );
+          XmppSender.sendPresence( xmppAccountRoot.xmppSession.xmlWriter, null, xmppItem.jid.concat( "/" ).concat( nickField.getText() ), null, XmppStatusUtil.getStatusDescr( (int)xmppAccountRoot.statusId ), null, 5, false, null, null );
           xmppItem.groupChatNick = nickField.getText();
         } catch ( IOException ex ) {
           ex.printStackTrace();
